requirements.txt
Django>=3.0.8,<3.1
djangorestframework>=3.11.0,<3.12
flake8>=3.8.3,<3.9
psycopg2>=2.7.5,<2.8.0
coveralls>=2.1.1,<2.2
